Init work on loan policies
